Refactor Teleport exports
Revert changes to export methods so future upgrades don't have as many
merge conflicts. Instead, create a new exported method of the same name
that calls the unexported method.
